+++ script/tool/test/pubspec_check_command_test.dart
@@ -146,15 +146,27 @@ void main() {
     });
 
     test('passes for a plugin following conventions', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin = createFakePlugin('plugin', packagesDir);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true)}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
 ${_dependenciesSection()}
 ${_devDependenciesSection()}
 ${_falseSecretsSection()}
+''');
+
+      plugin.getExamples().first.pubspecFile.writeAsStringSync('''
+${_headerSection(
+        'plugin_example',
+        publishable: false,
+        includeRepository: false,
+        includeIssueTracker: false,
+      )}
+${_environmentSection()}
+${_dependenciesSection()}
+${_flutterSection()}
 ''');
 
       final List<String> output = await runCapturingPrint(runner, <String>[
@@ -172,15 +184,28 @@ ${_falseSecretsSection()}
     });
 
     test('passes for a Flutter package following conventions', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage package =
+          createFakePackage('a_package', packagesDir);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${_headerSection('plugin')}
+      package.pubspecFile.writeAsStringSync('''
+${_headerSection('a_package')}
 ${_environmentSection()}
 ${_dependenciesSection()}
 ${_devDependenciesSection()}
 ${_flutterSection()}
 ${_falseSecretsSection()}
+''');
+
+      package.getExamples().first.pubspecFile.writeAsStringSync('''
+${_headerSection(
+        'a_package',
+        publishable: false,
+        includeRepository: false,
+        includeIssueTracker: false,
+      )}
+${_environmentSection()}
+${_dependenciesSection()}
+${_flutterSection()}
 ''');
 
       final List<String> output = await runCapturingPrint(runner, <String>[
@@ -190,18 +215,18 @@ ${_falseSecretsSection()}
       expect(
         output,
         containsAllInOrder(<Matcher>[
-          contains('Running for plugin...'),
-          contains('Running for plugin/example...'),
+          contains('Running for a_package...'),
+          contains('Running for a_package/example...'),
           contains('No issues found!'),
         ]),
       );
     });
 
     test('passes for a minimal package following conventions', () async {
-      final Directory packageDirectory = packagesDir.childDirectory('package');
-      packageDirectory.createSync(recursive: true);
+      final RepositoryPackage package =
+          createFakePackage('package', packagesDir, examples: <String>[]);
 
-      packageDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      package.pubspecFile.writeAsStringSync('''
 ${_headerSection('package')}
 ${_environmentSection()}
 ${_dependenciesSection()}
@@ -221,9 +246,10 @@ ${_dependenciesSection()}
     });
 
     test('fails when homepage is included', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, includeHomepage: true)}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
@@ -248,9 +274,10 @@ ${_devDependenciesSection()}
     });
 
     test('fails when repository is missing', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, includeRepository: false)}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
@@ -274,9 +301,10 @@ ${_devDependenciesSection()}
     });
 
     test('fails when homepage is given instead of repository', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, includeHomepage: true, includeRepository: false)}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
@@ -301,9 +329,10 @@ ${_devDependenciesSection()}
     });
 
     test('fails when repository is incorrect', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, repositoryPackagesDirRelativePath: 'different_plugin')}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
@@ -327,9 +356,10 @@ ${_devDependenciesSection()}
     });
 
     test('fails when issue tracker is missing', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, includeIssueTracker: false)}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
@@ -353,10 +383,11 @@ ${_devDependenciesSection()}
     });
 
     test('fails when description is too short', () async {
-      final Directory pluginDirectory =
-          createFakePlugin('a_plugin', packagesDir.childDirectory('a_plugin'));
+      final RepositoryPackage plugin = createFakePlugin(
+          'a_plugin', packagesDir.childDirectory('a_plugin'),
+          examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, description: 'Too short')}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
@@ -383,9 +414,10 @@ ${_devDependenciesSection()}
     test(
         'allows short descriptions for non-app-facing parts of federated plugins',
         () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, description: 'Too short')}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
@@ -410,14 +442,15 @@ ${_devDependenciesSection()}
     });
 
     test('fails when description is too long', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       const String description = 'This description is too long. It just goes '
           'on and on and on and on and on. pub.dev will down-score it because '
           'there is just too much here. Someone shoul really cut this down to just '
           'the core description so that search results are more useful and the '
           'package does not lose pub points.';
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, description: description)}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
@@ -442,9 +475,10 @@ ${_devDependenciesSection()}
     });
 
     test('fails when environment section is out of order', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true)}
 ${_flutterSection(isPlugin: true)}
 ${_dependenciesSection()}
@@ -469,9 +503,10 @@ ${_environmentSection()}
     });
 
     test('fails when flutter section is out of order', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true)}
 ${_flutterSection(isPlugin: true)}
 ${_environmentSection()}
@@ -496,9 +531,10 @@ ${_devDependenciesSection()}
     });
 
     test('fails when dependencies section is out of order', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true)}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
@@ -523,9 +559,9 @@ ${_dependenciesSection()}
     });
 
     test('fails when dev_dependencies section is out of order', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin = createFakePlugin('plugin', packagesDir);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true)}
 ${_environmentSection()}
 ${_devDependenciesSection()}
@@ -550,9 +586,10 @@ ${_dependenciesSection()}
     });
 
     test('fails when false_secrets section is out of order', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true)}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
@@ -579,10 +616,11 @@ ${_devDependenciesSection()}
 
     test('fails when an implemenation package is missing "implements"',
         () async {
-      final Directory pluginDirectory = createFakePlugin(
-          'plugin_a_foo', packagesDir.childDirectory('plugin_a'));
+      final RepositoryPackage plugin = createFakePlugin(
+          'plugin_a_foo', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin_a_foo', isPlugin: true)}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true)}
@@ -607,10 +645,11 @@ ${_devDependenciesSection()}
 
     test('fails when an implemenation package has the wrong "implements"',
         () async {
-      final Directory pluginDirectory = createFakePlugin(
-          'plugin_a_foo', packagesDir.childDirectory('plugin_a'));
+      final RepositoryPackage plugin = createFakePlugin(
+          'plugin_a_foo', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin_a_foo', isPlugin: true)}
 ${_environmentSection()}
 ${_flutterSection(isPlugin: true, implementedPackage: 'plugin_a_foo')}
@@ -635,10 +674,11 @@ ${_devDependenciesSection()}
     });
 
     test('passes for a correct implemenation package', () async {
-      final Directory pluginDirectory = createFakePlugin(
-          'plugin_a_foo', packagesDir.childDirectory('plugin_a'));
+      final RepositoryPackage plugin = createFakePlugin(
+          'plugin_a_foo', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection(
         'plugin_a_foo',
         isPlugin: true,
@@ -663,10 +703,11 @@ ${_devDependenciesSection()}
     });
 
     test('fails when a "default_package" looks incorrect', () async {
-      final Directory pluginDirectory =
-          createFakePlugin('plugin_a', packagesDir.childDirectory('plugin_a'));
+      final RepositoryPackage plugin = createFakePlugin(
+          'plugin_a', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection(
         'plugin_a',
         isPlugin: true,
@@ -702,10 +743,11 @@ ${_devDependenciesSection()}
     test(
         'fails when a "default_package" does not have a corresponding dependency',
         () async {
-      final Directory pluginDirectory =
-          createFakePlugin('plugin_a', packagesDir.childDirectory('plugin_a'));
+      final RepositoryPackage plugin = createFakePlugin(
+          'plugin_a', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection(
         'plugin_a',
         isPlugin: true,
@@ -739,10 +781,11 @@ ${_devDependenciesSection()}
     });
 
     test('passes for an app-facing package without "implements"', () async {
-      final Directory pluginDirectory =
-          createFakePlugin('plugin_a', packagesDir.childDirectory('plugin_a'));
+      final RepositoryPackage plugin = createFakePlugin(
+          'plugin_a', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection(
         'plugin_a',
         isPlugin: true,
@@ -768,11 +811,11 @@ ${_devDependenciesSection()}
 
     test('passes for a platform interface package without "implements"',
         () async {
-      final Directory pluginDirectory = createFakePlugin(
-          'plugin_a_platform_interface',
-          packagesDir.childDirectory('plugin_a'));
+      final RepositoryPackage plugin = createFakePlugin(
+          'plugin_a_platform_interface', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection(
         'plugin_a_platform_interface',
         isPlugin: true,
@@ -798,12 +841,13 @@ ${_devDependenciesSection()}
     });
 
     test('validates some properties even for unpublished packages', () async {
-      final Directory pluginDirectory = createFakePlugin(
-          'plugin_a_foo', packagesDir.childDirectory('plugin_a'));
+      final RepositoryPackage plugin = createFakePlugin(
+          'plugin_a_foo', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
       // Environment section is in the wrong location.
       // Missing 'implements'.
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection('plugin_a_foo', isPlugin: true, publishable: false)}
 ${_flutterSection(isPlugin: true)}
 ${_dependenciesSection()}
@@ -829,11 +873,12 @@ ${_environmentSection()}
     });
 
     test('ignores some checks for unpublished packages', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       // Missing metadata that is only useful for published packages, such as
       // repository and issue tracker.
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      plugin.pubspecFile.writeAsStringSync('''
 ${_headerSection(
         'plugin',
         isPlugin: true,
@@ -885,10 +930,10 @@ ${_devDependenciesSection()}
     });
 
     test('repository check works', () async {
-      final Directory packageDirectory =
-          createFakePackage('package', packagesDir);
+      final RepositoryPackage package =
+          createFakePackage('package', packagesDir, examples: <String>[]);
 
-      packageDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+      package.pubspecFile.writeAsStringSync('''
 ${_headerSection('package')}
 ${_environmentSection()}
 ${_dependenciesSection()}